+++ examples/rust/exercises/part-3/src/lib.rs
@@ -24,7 +24,7 @@ mod logic; // TODO: Implement this module.
 mod serial;
 
 fn main() {
-    let uart = uart::Uart(0);
+    let uart = uart::Uart::new(0).unwrap();
     let mut serial = Serial::new(&uart);
     let mut logic = Logic::new(); // NEW
     loop {