+++ examples/rust/exercises/part-5/src/lib.rs
@@ -32,7 +32,7 @@ mod store;
 mod touch; // TODO: Implement this module.
 
 fn main() {
-    let uart = uart::Uart(0);
+    let uart = uart::Uart::new(0).unwrap();
     let mut serial = Serial::new(&uart);
     let mut logic = Logic::new();
     loop {