+++ examples/rust/exercises/part-6-sol/src/lib.rs
@@ -24,7 +24,7 @@ mod store;
 mod touch;
 
 fn main() {
-    let uart = uart::Uart(0);
+    let uart = uart::Uart::new(0).unwrap();
     let mut serial = Serial::new(&uart);
     let mut logic = Logic::new();
     loop {