+++ examples/rust/exercises/part-6/src/lib.rs
@@ -26,7 +26,7 @@ mod store;
 mod touch; // TODO: Update this module.
 
 fn main() {
-    let uart = uart::Uart(0);
+    let uart = uart::Uart::new(0).unwrap();
     let mut serial = Serial::new(&uart);
     let mut logic = Logic::new();
     loop {